added alternative options and handler registration
Removed the registration of options and handlers from the constructor.
Instead there are separate function to change an options of to register
a handler.

Because of this it was not possible to create the Collection object in
the constructor This is now done on the first log call. A, for my
liking, way to complicated thingy with an object locking mechanism was
needed to make sure that after something was logged the options cannot
be changed anymore.